Add pushchanges function and remove go routine
pushchanges adds changes to the staging area and commits them with a
specified message. This will eventually become a push -> pr function as
well so pushChanges can be considered a wip.

This commit also contains changes to remove the go routine for
concurrency. The routine wasn't returning an error so will be added at a
later date.