Split all internal git actions into packages
Having intended to use the package name `git`, it became evident that
`go-git` already (and rightly so) uses this package name. I instead used
the name `gitAction` to signify the action git is performing. Having
read the go blog post, I'm still unclear if this is an idiomatic go
package name and will change it in the future if proved otherwise.